modified text editor

+++ database/migrations/2018_02_19_030304_create_menu_translations_table.php
@@ -13,15 +13,17 @@ class CreateMenuTranslationsTable extends Migration
      */
     public function up()
     {
-        Schema::create('menu_translations', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('menu_id')->unsigned();
-            $table->string('locale')->index();
-            $table->string('title');
-            $table->unique(['menu_id','locale']);
-            $table->timestamps();
-            $table->foreign('menu_id')->references('id')->on('menus')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('menu_translations')) {
+            Schema::create('menu_translations', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('menu_id')->unsigned();
+                $table->string('locale')->index();
+                $table->string('title');
+                $table->unique(['menu_id','locale']);
+                $table->timestamps();
+                $table->foreign('menu_id')->references('id')->on('menus')->onDelete('cascade');
+            });
+        }
     }
 
     /**
